requirements.txt
aiohttp==3.8.5
aiosignal==1.2.0
altair==5.0.1
anyio==3.7.1
appnope==0.1.2
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
blinker==1.4
Bottleneck==1.3.5
brotlipy==0.7.0
cachetools==4.2.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
click==8.0.4
comm==0.1.2
cryptography==41.0.3
dataclasses-json==0.5.7
debugpy==1.6.7
decorator==5.1.1
docopt==0.6.2
exceptiongroup==1.0.4
executing==0.8.3
fastapi==0.103.1
frozenlist==1.3.3
gitdb==4.0.7
GitPython==3.1.30
google-api-core==2.12.0
google-auth==2.22.0
googleapis-common-protos==1.56.4
greenlet==2.0.1
h11==0.12.0
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.25.0
ipython==8.15.0
ipywidgets==8.0.4
jedi==0.18.1
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.1
jsonschema==4.17.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.5
langchain==0.0.305
langsmith==0.0.41
markdown-it-py==2.2.0
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mdurl==0.1.0
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
multidict==6.0.2
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numexpr==2.8.4
numpy==1.25.2
openai==0.28.1
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.2.1
pipreqs==0.4.13
platformdirs==3.10.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.8
pydeck==0.7.1
Pygments==2.15.1
Pympler==0.9
pyOpenSSL==23.2.0
PyPDF2==2.10.5
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0
pyzmq==25.1.0
requests==2.31.0
rich==13.3.5
rsa==4.7.2
setuptools==68.0.0
six==1.16.0
smmap==4.0.0
sniffio==1.2.0
SQLAlchemy==2.0.21
stack-data==0.2.0
starlette==0.27.0
streamlit==1.23.1
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.66.1
traitlets==5.7.1
typing_extensions==4.7.1
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==2.1
urllib3==1.26.16
uvicorn==0.20.0
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
wheel==0.38.4
widgetsnbextension==4.0.5
yarg==0.1.9
yarl==1.8.1
zipp==3.11.0

reportlab~=3.6.12